+++ packages/gatsby/src/schema/resolvers.js
@@ -78,7 +78,7 @@ const paginate = (results = [], { skip = 0, limit }) => {
   const hasNextPage = skip + limit < count
 
   return {
-    totalCount: items.length,
+    totalCount: count,
     edges: items.map((item, i, arr) => {
       return {
         node: item,
@@ -151,9 +151,7 @@ const fileByPath = ({ from }) => (source, args, context, info) => {
     return fieldValue
   }
 
-  const isArray = getNullableType(info.returnType) instanceof GraphQLList
-
-  const findLinkedFileNode = async relativePath => {
+  const findLinkedFileNode = relativePath => {
     // Use the parent File node to create the absolute path to
     // the linked file.
     const fileLinkPath = normalize(
@@ -162,7 +160,7 @@ const fileByPath = ({ from }) => (source, args, context, info) => {
 
     // Use that path to find the linked File node.
     const linkedFileNode = _.find(
-      await context.nodeModel.getAllNodes({ type: `File` }),
+      context.nodeModel.getAllNodes({ type: `File` }),
       n => n.absolutePath === fileLinkPath
     )
     return linkedFileNode
@@ -170,15 +168,18 @@ const fileByPath = ({ from }) => (source, args, context, info) => {
 
   // Find the File node for this node (we assume the node is something
   // like markdown which would be a child node of a File node).
-  const parentFileNode = context.nodeModel.findRootNodeAncestor(source)
+  const parentFileNode = context.nodeModel.findRootNodeAncestor(
+    source,
+    node => node.internal && node.internal.type === `File`
+  )
 
-  // Find the linked File node(s)
-  if (isArray) {
-    return Promise.all(fieldValue.map(findLinkedFileNode))
-  } else {
-    return findLinkedFileNode(fieldValue)
-  }
+  return resolveValue(findLinkedFileNode, fieldValue)
 }
+
+const resolveValue = (resolve, value) =>
+  Array.isArray(value)
+    ? value.map(v => resolveValue(resolve, v))
+    : resolve(value)
 
 module.exports = {
   findManyPaginated,